fix:  fixing header navigation && create auth pages
