+++ src/app/(auth)/forgot-password/page.tsx
@@ -1,0 +1,3 @@
+export default function ForgotPasswordPage() {
+  return <div>ForgotPasswordPage</div>
+}
