+++ src/app/(auth)/sign-in/page.tsx
@@ -1,0 +1,3 @@
+export default function SignIn() {
+  return <div>SignIn</div>
+}
