+++ src/app/(auth)/sign-up/page.tsx
@@ -1,0 +1,3 @@
+export default function SignUp() {
+  return <div>SignUp</div>
+}
