+++ src/app/layout.tsx
@@ -7,6 +7,8 @@ import RootHead from '@/components/helpers/RootHead'
 
 import './globals.css'
 
+import HeaderOrMenu from '@/components/common/Home/HeaderOrMenu/HeaderOrMenu'
+
 const CormorantFont = Cormorant({
   variable: '--font-cormorant',
   weight: ['500', '600', '400'],
@@ -20,7 +22,7 @@ const ManropeFont = Manrope({
 })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
+  title: 'Hypnos App',
   description: 'Generated by create next app'
 }
 
@@ -35,7 +37,10 @@ export default function RootLayout({
       <body
         className={`${ManropeFont.className} ${CormorantFont.variable} font-manrope text-black
           antialiased`}>
-        <Providers>{children}</Providers>
+        <Providers>
+          <HeaderOrMenu />
+          {children}
+        </Providers>
       </body>
     </html>
   )
