+++ src/components/common/Home/HeaderOrMenu/components/Header/HeaderPageNavigation.tsx
@@ -2,7 +2,7 @@
 
 import { Media, MediaContextProvider } from '@/components/helpers/Media'
 
-import TitleNavigationLink from './TitleNavigationLink'
+import HeaderPageNavigationLink from './HeaderPageNavigationLink'
 
 const navLinks = [
   { href: '#catalogue', label: 'Catalogue' },
@@ -11,14 +11,14 @@ const navLinks = [
   { href: '#contacts', label: 'Contacts' }
 ]
 
-export default function TitleNavigation() {
+export default function HeaderPageNavigation() {
   return (
     <MediaContextProvider>
       <Media greaterThanOrEqual='md'>
         <nav className='mb:mb-3 hidden md:block'>
           <ul className='flex justify-center gap-slide-large font-manrope text-sm text-grey-400'>
             {navLinks.map(({ href, label }) => (
-              <TitleNavigationLink
+              <HeaderPageNavigationLink
                 key={href}
                 href={href}
                 label={label}
