+++ src/components/modals/BurgerModal/data/burger-data.ts
@@ -1,16 +1,16 @@
 export const navItems = [
-    { to: '/', title: 'Home' },
-    { to: '/news', title: 'News' },
-    { to: '/aboutus', title: 'About us' },
-    { to: '/cataloge', title: 'Rings' },
-    { to: '/contacts', title: 'Contacts' }
-  ]
-  
+  { to: '/', title: 'Home' },
+  { to: '/news', title: 'News' },
+  { to: '/aboutus', title: 'About us' },
+  { to: '/cataloge', title: 'Rings' },
+  { to: '/contacts', title: 'Contacts' }
+]
+
 export const addressItems = [
-    {
-      to: 'https://maps.app.goo.gl/DcGGzwfaerAshAA69',
-      title: '123 Walnut Street, USA'
-    },
-    { to: 'mailto:webmaster@example.com', title: 'webmaster@example.com' },
-    { to: 'tel:+12155550199', title: '+12155550199' }
-  ]
+  {
+    to: 'https://maps.app.goo.gl/DcGGzwfaerAshAA69',
+    title: '123 Walnut Street, USA'
+  },
+  { to: 'mailto:webmaster@example.com', title: 'webmaster@example.com' },
+  { to: 'tel:+12155550199', title: '+12155550199' }
+]
